Module oo7::dbus

The implementation in this module allows you to interact with a org.freedesktop.Secrets implementation.

That is usually done with

let service = Service::new(Algorithm::Plain).await?;

let mut attributes = std::collections::HashMap::new();
attributes.insert("type", "password");
attributes.insert("user_id", "some_other_identifier");

if let Some(collection) = service.default_collection().await? {
    // Store a secret
    collection
        .create_item(
            "My App's secret",
            attributes.clone(),
            b"password",
            true,
            "text/plain",
        )
        .await?;

    // Retrieve it later thanks to it attributes
    let item = collection.search_items(attributes).await?.first().unwrap();
    assert_eq!(item.secret().await?, b"password");
}

Modules

api

Barebone DBus API of the secret service specifications.

Structs

Collection

A collection allows to store and retrieve items.

Item

A secret with a label and attributes to identify it.

Service

The entry point of communicating with a org.freedesktop.Secrets implementation.

Enums

Algorithm

Statics

DEFAULT_COLLECTION

The default collection alias.

SESSION_COLLECTION

A session collection.