Struct oo7::dbus::Item

pub struct Item<'a> { /* private fields */ }

A secret with a label and attributes to identify it.

An item might be locked or unlocked, use Item::lock or Item::unlock to lock or unlock it. Note that the secret service might not be able to lock/unlock individual items and may lock/unlock the entire collection in such case.

The item is attributes are used to identify and find the item later using Collection::search_items. They are not stored or transferred in a secure manner.

Note

If the item is deleted using Item::delete any future usage of it API will fail with Error::Deleted.

Implementations

impl<'a> Item<'a>

pub async fn is_locked(&self) -> Result<bool, Error>

Get whether the item is locked.

pub async fn label(&self) -> Result<String, Error>

The item label.

pub async fn set_label(&self, label: &str) -> Result<(), Error>

Set the item label.

pub async fn created(&self) -> Result<Duration, Error>

The UNIX time when the item was created.

pub async fn modified(&self) -> Result<Duration, Error>

The UNIX time when the item was modified.

pub async fn attributes(&self) -> Result<HashMap<String, String>, Error>

Retrieve the item attributes.

pub async fn set_attributes(
&self,
attributes: HashMap<&str, &str>
) -> Result<(), Error>

Update the item attributes.

pub async fn delete(&self) -> Result<(), Error>

Delete the item.

pub async fn secret(&self) -> Result<Zeroizing<Vec<u8>>, Error>

Retrieve the currently stored secret.

pub async fn set_secret<P: AsRef<[u8]>>(
&self,
secret: P,
content_type: &str
) -> Result<(), Error>

Modify the stored secret on the item.

Arguments

secret - The secret to store.

content_type - The content type of the secret, usually something like text/plain.

pub async fn unlock(&self) -> Result<(), Error>

Unlock the item.

pub async fn lock(&self) -> Result<(), Error>

Lock the item.